refactor: improve folder structure for @lifeforge/configs
