+++ packages/configs/src/index.ts
@@ -1,3 +1,7 @@
-export { defineModuleConfig } from './module-vite.config'
-export { defineClientConfig, SHARED_PACKAGES } from './client-vite.config'
-export { customResolver } from './resolver'
+export { defineModuleConfig } from './vite/module-vite.config'
+
+export { defineClientConfig } from './vite/client-vite.config'
+
+export { SHARED_PACKAGES } from './constants/shared-packages'
+
+export { aliasResolver as customResolver } from './resolvers/alias-resolver'
