+++ packages/configs/src/resolvers/alias-resolver.ts
@@ -6,21 +6,28 @@ import path from 'node:path'
  * Supports resolving from the client src directory, client root for manifest files,
  * and falls back to package src directories.
  */
-export function customResolver(id: string, importer: string | undefined): string | null {
+export function aliasResolver(
+  id: string,
+  importer: string | undefined
+): string | null {
   if (!importer) {
     return null
   }
 
   let rootDir: string
+
   const clientMatch = importer.match(/(.+\/client)/)
+
   if (clientMatch) {
     const clientDir = clientMatch[1]
+
     rootDir =
       id === '@/manifest' || id === '@/manifest.ts'
         ? clientDir
         : `${clientDir}/src`
   } else {
     const srcMatch = importer.match(/(.+\/src)/)
+
     if (!srcMatch) {
       return null
     }
@@ -28,6 +35,7 @@ export function customResolver(id: string, importer: string | undefined): string
   }
 
   const basePath = path.resolve(rootDir, id.slice(2))
+
   const candidates = [
     `${basePath}.tsx`,
     `${basePath}.ts`,
@@ -43,8 +51,7 @@ export function customResolver(id: string, importer: string | undefined): string
     }
   }
 
-  console.log(
-    `[vite] failed to resolve import "${id}" from "${importer}"`
-  )
+  console.log(`[vite] failed to resolve import "${id}" from "${importer}"`)
+
   return null
 }
