+++ packages/configs/src/vite/module-vite.config.ts
@@ -3,6 +3,8 @@ import react from '@vitejs/plugin-react'
 import path from 'node:path'
 import { type UserConfig, defineConfig, loadEnv, mergeConfig } from 'vite'
 
+import { SHARED_PACKAGES } from '../constants/shared-packages'
+
 interface ModuleConfigOptions {
   dirname: string
   pkg: {
@@ -20,15 +22,19 @@ export function defineModuleConfig(
   overrides: UserConfig | ((env: any) => UserConfig) = {}
 ) {
   const isDocker = process.env.DOCKER_BUILD === 'true'
+
   const outDir = isDocker ? 'dist-docker' : 'dist'
+
   const moduleName = pkg.name.replace('@lifeforge/', '')
 
-  return defineConfig((configEnv) => {
+  return defineConfig(configEnv => {
     const isDev = configEnv.command === 'serve'
 
     // 1. Resolve environment variables natively from the root env folder
     const envDir = path.resolve(dirname, '../../../env')
+
     const env = loadEnv(configEnv.mode, envDir, '')
+
     const apiHost = isDocker ? '/api' : env.VITE_API_HOST
 
     if (!apiHost) {
@@ -50,38 +56,24 @@ export function defineModuleConfig(
           shared: {
             react: { generate: false },
             'react-dom': { generate: false },
-            '@lifeforge/shared': { generate: false },
-            '@lifeforge/federation': { generate: false },
-            '@lifeforge/ui': { generate: false },
             'react-i18next': { generate: false },
             i18next: { generate: false },
-            '@tanstack/react-query': { generate: false }
+            '@tanstack/react-query': { generate: false },
+            ...Object.fromEntries(
+              Object.keys(SHARED_PACKAGES).map(e => [e, { generate: false }])
+            )
           }
         })
       ],
       resolve: {
         alias: [
           ...(isDev
-            ? [
-                {
-                  find: /^@lifeforge\/shared$/,
-                  replacement: path.resolve(dirname, '../../../shared/src/index.ts')
-                },
-                {
-                  find: /^@lifeforge\/federation$/,
-                  replacement: path.resolve(
-                    dirname,
-                    '../../../packages/federation/src/index.ts'
-                  )
-                },
-                {
-                  find: /^@lifeforge\/ui$/,
-                  replacement: path.resolve(
-                    dirname,
-                    '../../../packages/ui/src/index.ts'
-                  )
-                }
-              ]
+            ? Object.entries(SHARED_PACKAGES).map(
+                ([, { aliasRegex, entryPoint }]) => ({
+                  find: aliasRegex,
+                  replacement: path.resolve(dirname, '/../../', entryPoint)
+                })
+              )
             : []),
           {
             find: /^@\/manifest$/,
@@ -99,9 +91,8 @@ export function defineModuleConfig(
     }
 
     // 3. Resolve custom overrides if they are passed as a function
-    const resolvedOverrides = typeof overrides === 'function' 
-      ? overrides(configEnv) 
-      : overrides
+    const resolvedOverrides =
+      typeof overrides === 'function' ? overrides(configEnv) : overrides
 
     // 4. Merge base configuration with the module-specific overrides
     return mergeConfig(baseConfig, resolvedOverrides)
